Core: merge multiple Contract methods, using INumber<T> whenever possible
- merge all int/long/double/etc.. overloads to use INumber<T> for comparisons and modulo
- on .NET 6, we have to fallback to IComparable<T> and keep the Multiple(...) overloads
- using IComparable<T> seems to produce a double bound check with current JIT
- we only lose suport for structs like TimeSpan, but this is easy to change

+++ Doxense.Core/Memory/MutableSlice.cs
@@ -116,7 +116,7 @@ namespace System
 		[Pure]
 		public static MutableSlice Zero(uint size)
 		{
-			Contract.LessOrEqual(size, int.MaxValue);
+			Contract.LessOrEqual(size, (uint) int.MaxValue);
 			return size != 0 ? new MutableSlice(new byte[size], 0, (int) size) : MutableSlice.Empty;
 		}
